requirements/base.txt
Django==1.11.4
django-extensions==1.8.1
django-model-utils==3.0.0
djangorestframework==3.6.3
django-js-reverse==0.7.3
psycopg2==2.7.3